Update server to run c++ witness generation

+++ index.js
@@ -1,5 +1,4 @@
 const express = require("express");
-const snarkjs = require("snarkjs");
 const hash = require("object-hash");
 const fs = require("fs");
 const { spawn } = require("child_process");
@@ -37,9 +36,11 @@ const inQueue = (hash) => {
 const startNewProcess = (hash) => {
   if (!!outputData[hash]) return;
   const inputFileName = `inputs/${hash}.json`;
-  const witnessFileName = `inputs/${hash}.json_witness`;
+  const witnessFileName = `inputs/${hash}.wtns`;
+  const publicName = `public_${hash}.json`;
+  const proofName = `proof_${hash}.json`;
   // spawn a child process to run the proof generation
-  const prover = spawn("node", ["prover.js", inputFileName], {
+  const prover = spawn("sh", ["exec.sh", hash], {
     timeout: 60 * 60 * 1000,
   });
   if (!prover.pid) {
@@ -47,34 +48,41 @@ const startNewProcess = (hash) => {
     return false;
   }
   currentProcessesRunning.add(hash);
-  prover.stderr.on("data", (data) => {
-    outputData[hash] = data.toString();
-    console.log(`stderr: ${hash} :  ${data}`);
-    console.log("outputData", outputData);
+  prover.stdout.on("data", (data) => {
+    var res = data.toString();
+    if (res.substring(0,8) === "{\"pi_a\":") {
+        outputData[hash] = res;
+        // delete the relevant files in inputs folder
+        fs.unlinkSync(inputFileName);
+        fs.unlinkSync(witnessFileName);
+        fs.unlinkSync(publicName);
+        fs.unlinkSync(proofName);
+    } else {
+        outputData[hash] = '{\"result\": \"BLS signature verification failed.\"}';
+    }
     currentProcessesRunning.delete(hash);
     processQueue();
     prover.kill();
-    // delete the relevant files in inputs folder
-    fs.unlinkSync(inputFileName);
-    fs.unlinkSync(witnessFileName);
   });
 
-  prover.stdout.on("data", (data) => {
-    console.error(`stdout: ${data}`);
+  prover.stderr.on("data", (data) => {
+    console.error(`stderr: ${data}`);
+    outputData[hash] = '{\"result\": \"BLS signature verification failed.\"}';
+    currentProcessesRunning.delete(hash);
+    processQueue();
+    prover.kill();
   });
 
   prover.on("close", (code) => {
     currentProcessesRunning.delete(hash);
     processQueue();
     console.log(`child process exited with code ${code}`);
-//    fs.unlinkSync(inputFileName);
-//    fs.unlinkSync(witnessFileName);
   });
   return true;
 };
 
 const processQueue = () => {
-  if (currentProcessesRunning.size >= 2) return null;
+  if (currentProcessesRunning.size >= 1) return null;
 
   // remove duplicates from queue
   queue = queue.filter(
@@ -126,21 +134,22 @@ app.post("/result", (req, res) => {
 
   if (result) {
     try {
+      console.log(result);
       JSON.parse(result);
       res.send(result);
     } catch (e) {
       console.log("error", e);
-      res.status(404).send("ERROR");
+      res.status(404).send("{\"result\": \"ERROR\"}");
     }
   } else if (currentProcessesRunning.has(id) || inQueue(id)) {
     console.log("waiting for result");
     console.log("currentProcessesRunning", currentProcessesRunning);
-    res.status(400).send("Process still running");
+    res.status(400).send("{\"result\": \"Process still running\"}");
   } else {
     console.log("ERROR! result not found", id, queue);
     // print outputdata keys
     console.log("outputData", outputData);
-    res.status(404).send("ERROR");
+    res.status(404).send("{\"result\": \"ERROR\"}");
   }
 });
